Password Grader
This program uses different variables to assign a "grade" to any given password. It factors in length, uppercase letters, numbers, and special characters into the creation of this grade. It also uses a list of the 100k most popular passwords taken from the internet to make sure that the password is not too easy to guess. The program uses tkinter for UI/UX.